+++ src/context/contextview.cpp
@@ -498,7 +498,7 @@ void ContextView::UpdateSong(const Song &song) {
 
 void ContextView::UpdateLyrics(const quint64 id, const QString &provider, const QString &lyrics) {
 
-  if (id != lyrics_id_) return;
+  if ((qint64) id != lyrics_id_) return;
   lyrics_ = lyrics + "\n\n(Lyrics from " + provider + ")\n";
   lyrics_id_ = -1;
   if (action_show_lyrics_->isChecked()) {
